refactor(cockpit): extract _InboxListItem into sibling module (#1354)
Ninth wedge of the cockpit_ui.py god-module split. Moves the
``_InboxListItem`` widget (the inbox list-row ListItem subclass
carrying ``task_id``/``task_ref``/``is_unread`` + the plan-review
/ rejection-feedback / triage-bucket class hooks for #1400, #1573)
into a new ``cockpit_inbox_list_item.py`` sibling and re-exports
it from ``cockpit_ui`` for back-compat (tests + the inbox app's
isinstance checks still reach it by the original name).

The ``_format_inbox_thread_row`` / ``_is_plan_review_task`` /
``_triage_bucket`` helpers still live in ``cockpit_ui``; the new
module lazy-imports them from inside ``__init__`` /
``mark_read`` / ``set_show_judgment_calls`` so the load-time
cycle (cockpit_ui imports the sibling for the re-export shim)
stays broken — same pattern the sixth wedge (#1700,
``_RollupItem``) and earlier extractions used.

LOC progression for ``cockpit_ui.py``:
- Pre-wedge: 18,620
- Post-wedge: 18,543 (-77)

Inbox + project-dashboard test sweep stays green (554 passed via
``pytest -k inbox``).
